% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloderiv_functions.R
\name{compute.phyderiv.intvl}
\alias{compute.phyderiv.intvl}
\title{Compute the phylogenetic derivative of a family of trees using the path interval distance.}
\usage{
compute.phyderiv.intvl(trees)
}
\arguments{
\item{trees}{a list of phylogenetic trees.}
}
\value{
a vector with the phylogenetic derivative
}
\description{
Compute the phylogenetic derivative of a family of trees using the path interval distance.
}
