% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloderiv_functions.R
\name{compute.phyderiv.rf}
\alias{compute.phyderiv.rf}
\title{Compute the phylogenetic derivative of a family of trees using Robinson-Foulds distance.}
\usage{
compute.phyderiv.rf(trees)
}
\arguments{
\item{trees}{a list of phylogenetic trees.}
}
\value{
a vector with the phylogenetic derivative
}
\description{
Compute the phylogenetic derivative of a family of trees using Robinson-Foulds distance.
}
