% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloderiv_functions.R
\name{compute.treedists.intvl}
\alias{compute.treedists.intvl}
\title{Compute the path interval distance between pairs of trees at corresponding indices.}
\usage{
compute.treedists.intvl(est, true)
}
\arguments{
\item{est}{a list of estimated trees}

\item{true}{a list of true trees}
}
\value{
a vector of path interval distances between corresponding true and est trees.
}
\description{
Compute the path interval distance between pairs of trees at corresponding indices.
}
