% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloderiv_functions.R
\name{compute.treedists.rf}
\alias{compute.treedists.rf}
\title{Compute the Robinson-Foulds distance between pairs of trees at corresponding indices.}
\usage{
compute.treedists.rf(est, true)
}
\arguments{
\item{est}{a list of estimated trees}

\item{true}{a list of true trees}
}
\value{
a vector of RF distances between corresponding true and est trees.
}
\description{
Compute the Robinson-Foulds distance between pairs of trees at corresponding indices.
}
