% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloderiv_functions.R
\name{readBlossoc}
\alias{readBlossoc}
\title{Read in the Blossoc trees}
\usage{
readBlossoc(treefile, indexfile)
}
\arguments{
\item{treefile}{the path and file name of the file with the qb trees (type string)}

\item{indexfile}{the path and file name of the file with the index (type string)}
}
\value{
a list of blossoc trees
}
\description{
Read in the Blossoc trees
}
