% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloderiv_functions.R
\name{readRent}
\alias{readRent}
\title{Read in the Rent trees}
\usage{
readRent(treefile)
}
\arguments{
\item{treefile}{the path and file name of the file with the rent trees (type string)}
}
\value{
a list of blossoc trees
}
\description{
Read in the Rent trees
}
